% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval-psi.R
\name{build_eval_psi_fun}
\alias{build_eval_psi_fun}
\title{Build Fast ψ-Conditional Optimizer for Branch Evaluation (Internal)}
\usage{
build_eval_psi_fun(cal)
}
\arguments{
\item{cal}{A \code{likelyr_calibrated} model object.}
}
\value{
A nested function:

\preformatted{
  function(omega_hat) {
    function(psi_target, param_init) { ... }
  }
}
}
\description{
Constructs a \emph{two-stage closure system} for fast repeated solutions:

\deqn{
  \param^{\*}(\psi, \omega)
    = \arg\max_{\param} E_{\omega}[ \log L(\param) ]
    \quad\text{s.t.}\quad
    ψ(\param)=ψ_{\mathrm{target}}.
}

Usage:

\preformatted{
f1 <- build_eval_psi_fun(cal)
f2 <- f1(omega_hat)
out <- f2(psi_target, param_init)
}
}
\keyword{internal}
