% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-diagnose-integrate-plots.R
\name{build_integrate_ess_plot}
\alias{build_integrate_ess_plot}
\title{Plot Effective Sample Size (ESS)}
\usage{
build_integrate_ess_plot(ess)
}
\description{
Plot Effective Sample Size (ESS)
}
