% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-diagnose-integrate-plots.R
\name{build_integrate_omega_branch_plot}
\alias{build_integrate_omega_branch_plot}
\title{Plot Omega-hat Log-Likelihood Branches}
\usage{
build_integrate_omega_branch_plot(branch_mat, psi_vals)
}
\description{
Plot Omega-hat Log-Likelihood Branches
}
