% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-diagnose-integrate-plots.R
\name{build_integrate_omega_eigen_plot}
\alias{build_integrate_omega_eigen_plot}
\title{Plot Omega-hat Covariance Eigenvalues}
\usage{
build_integrate_omega_eigen_plot(eigenvalues)
}
\description{
Plot Omega-hat Covariance Eigenvalues
}
