% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-diagnose-integrate-plots.R
\name{build_integrate_omega_pca_plot}
\alias{build_integrate_omega_pca_plot}
\title{Plot Omega-hat PCA Scatter}
\usage{
build_integrate_omega_pca_plot(Omega)
}
\description{
Plot Omega-hat PCA Scatter
}
