% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-diagnose-integrate-plots.R
\name{build_integrate_outlier_plot}
\alias{build_integrate_outlier_plot}
\title{Plot Outlier Fraction}
\usage{
build_integrate_outlier_plot(outlier_fraction)
}
\description{
Plot Outlier Fraction
}
