% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-calibrate.R
\name{calibrate_IL}
\alias{calibrate_IL}
\title{Calibrate a Workflow for Integrated Likelihood}
\usage{
calibrate_IL(wf, data, verbose = TRUE)
}
\arguments{
\item{wf}{A \code{likelihood_workflow} object created by \code{\link[=workflow]{workflow()}}.}

\item{data}{User data to embed into model/estimand closures.}

\item{verbose}{Logical; prints diagnostics if TRUE.}
}
\value{
\verb{"IL_calibration", "likelihood_calibration"}
}
\description{
Adds IL-specific calibration components:
\itemize{
\item equality constraint ψ(θ) = ψ_MLE,
\item nuisance initialization generator,
\item nuisance manifold sampler.
}
}
