% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-calibrate.R
\name{calibrate_PL}
\alias{calibrate_PL}
\title{Calibrate a Workflow for Profile Likelihood}
\usage{
calibrate_PL(wf, data, verbose = TRUE)
}
\arguments{
\item{wf}{A \code{likelihood_workflow} object created by \code{\link[=workflow]{workflow()}}.}

\item{data}{User data to embed into model/estimand closures.}

\item{verbose}{Logical; prints diagnostics if TRUE.}
}
\value{
\verb{"PL_calibration", "likelihood_calibration"}
}
\description{
Installs PL-specific calibration components (currently minimal but
extensible for future profile-SE functionality).
}
