% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-diagnose-integrate-metrics.R
\name{compute_integrate_ess}
\alias{compute_integrate_ess}
\title{Effective sample size (ESS) for integrated likelihood}
\usage{
compute_integrate_ess(cv2, R)
}
\arguments{
\item{cv2}{Numeric vector of squared coefficients of variation.}

\item{R}{Integer number of Monte Carlo draws.}
}
\value{
Numeric vector of effective sample sizes.
}
\description{
Computes effective sample size using the coefficient of variation
of Monte Carlo likelihood estimates.
}
\keyword{internal}
