% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ll-processing.R
\name{compute_MLE}
\alias{compute_MLE}
\title{Compute the MLE and maximum log-likelihood from a spline model}
\usage{
compute_MLE(spline_model, df)
}
\arguments{
\item{spline_model}{A smooth.spline object from fit_ll_spline().}

\item{df}{A data frame with psi and loglik.}
}
\value{
A tibble with MLE and Maximum.
}
\description{
Compute the MLE and maximum log-likelihood from a spline model
}
