% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-diagnose.R
\name{.diagnose_integrate_result}
\alias{.diagnose_integrate_result}
\title{Integrated log-likelihood diagnostics (full)}
\usage{
.diagnose_integrate_result(res)
}
\description{
Integrated log-likelihood diagnostics (full)
}
\keyword{internal}
