% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-diagnose.R
\name{.diagnose_integrated_result}
\alias{.diagnose_integrated_result}
\title{Internal integrated log-likelihood diagnostics engine}
\usage{
.diagnose_integrated_result(res)
}
\description{
Internal integrated log-likelihood diagnostics engine
}
\keyword{internal}
