% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-diagnose.R
\name{.diagnose_pl_object}
\alias{.diagnose_pl_object}
\title{Placeholder diagnostics for Profile Log-Likelihood}
\usage{
.diagnose_pl_object(pl)
}
\description{
Placeholder diagnostics for Profile Log-Likelihood
}
\keyword{internal}
