% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-diagnose.R
\name{.diagnose_profile_result}
\alias{.diagnose_profile_result}
\title{Placeholder diagnostics for profile log-likelihood}
\usage{
.diagnose_profile_result(res)
}
\description{
Placeholder diagnostics for profile log-likelihood
}
\keyword{internal}
