% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-style.R
\name{.get_comparison_value}
\alias{.get_comparison_value}
\title{Comparison-aware style lookup}
\usage{
.get_comparison_value(base_path, key, comparison)
}
\arguments{
\item{base_path}{Base node}

\item{key}{Field name}

\item{comparison}{Logical}
}
\value{
Style value
}
\description{
Comparison-aware style lookup
}
\keyword{internal}
