% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-style.R
\name{.get_style_node}
\alias{.get_style_node}
\title{Navigate nested style node}
\usage{
.get_style_node(path)
}
\arguments{
\item{path}{Character vector of node names}
}
\value{
Node value or NULL
}
\description{
Navigate nested style node
}
\keyword{internal}
