% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-style.R
\name{.get_style_value}
\alias{.get_style_value}
\title{Lookup style value with validation}
\usage{
.get_style_value(path, key, label)
}
\arguments{
\item{path}{Base path}

\item{key}{Value name}

\item{label}{Human-readable label for errors}
}
\value{
Requested style value
}
\description{
Lookup style value with validation
}
\keyword{internal}
