% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-style.R
\name{.resolve_pseudolikelihood_value}
\alias{.resolve_pseudolikelihood_value}
\title{Resolve scalar vs named list values}
\usage{
.resolve_pseudolikelihood_value(val, pseudolikelihood)
}
\description{
Resolve scalar vs named list values
}
\keyword{internal}
