% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ll-processing.R
\name{fit_ll_spline}
\alias{fit_ll_spline}
\title{Fit a spline model to a single log-likelihood curve}
\usage{
fit_ll_spline(df)
}
\arguments{
\item{df}{A data frame with columns psi and loglik.}
}
\value{
A smooth.spline object.
}
\description{
Fit a spline model to a single log-likelihood curve
}
