% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-palette.R
\name{flatten_table_palette}
\alias{flatten_table_palette}
\title{Flatten table palette to a named list}
\usage{
flatten_table_palette()
}
\description{
Useful for legacy code expecting pal$psi_hat, etc.
}
