% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-generate.R
\name{generate.PL_calibration}
\alias{generate.PL_calibration}
\title{Generate Profile Likelihood (not yet implemented)}
\usage{
\method{generate}{PL_calibration}(cal, verbose = TRUE, ...)
}
\arguments{
\item{cal}{A \code{"PL_calibration"} object.}

\item{verbose}{Logical; print progress messages.}

\item{...}{Ignored.}
}
\value{
Currently throws an error.
}
\description{
Placeholder method for profile likelihood generation.
Will eventually mirror the IL branch-generation framework.
}
