% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plots.R
\name{get_branches_plot}
\alias{get_branches_plot}
\title{Plot Integrated Log-Likelihood Branches}
\usage{
get_branches_plot(mat)
}
\arguments{
\item{mat}{A matrix of log-likelihood branches.}
}
\value{
A ggplot object.
}
\description{
Plot Integrated Log-Likelihood Branches
}
