% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plots.R
\name{get_branches_plot_shifted}
\alias{get_branches_plot_shifted}
\title{Plot Shifted Integrated Log-Likelihood Branches}
\usage{
get_branches_plot_shifted(mat, crit)
}
\arguments{
\item{mat}{Branch matrix.}

\item{crit}{Critical value for horizontal cutoff line.}
}
\value{
A ggplot object.
}
\description{
Plot Shifted Integrated Log-Likelihood Branches
}
