% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plots.R
\name{get_ci_colors}
\alias{get_ci_colors}
\title{Colors for Likelihood-Ratio Confidence Levels}
\usage{
get_ci_colors(crit_df)
}
\arguments{
\item{crit_df}{Data frame containing CI labels.}
}
\value{
Named vector of colors.
}
\description{
Colors for Likelihood-Ratio Confidence Levels
}
\keyword{internal}
