% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-palette.R
\name{get_ci_palette}
\alias{get_ci_palette}
\title{Colors for Likelihood-Ratio Confidence Levels}
\usage{
get_ci_palette(conf_ints)
}
\arguments{
\item{conf_ints}{Data frame of CI values.}
}
\value{
Named vector of colors for labelling CI endpoints on likelihood plot.
}
\description{
Colors for Likelihood-Ratio Confidence Levels
}
\keyword{internal}
