% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plots.R
\name{get_LL_plot}
\alias{get_LL_plot}
\title{Plot Log-Likelihood Curve}
\usage{
get_LL_plot(df)
}
\arguments{
\item{df}{A data frame with columns \code{psi} and a likelihood column.}
}
\value{
A ggplot object.
}
\description{
Plot Log-Likelihood Curve
}
