% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer-point_estimate.R
\name{get_psi_loglik_max_point}
\alias{get_psi_loglik_max_point}
\title{Locate the maximizer of a smoothed ψ log-likelihood}
\usage{
get_psi_loglik_max_point(psi_loglik)
}
\arguments{
\item{psi_loglik}{A function of the form returned by \code{fit_psi_loglik()}.}
}
\value{
A named numeric vector with elements \code{argmax} and \code{maximum}.
}
\description{
Finds the maximizer and maximum value of a smoothed log-likelihood
function over the ψ grid range.
}
\keyword{internal}
