% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plots.R
\name{get_stat_fns}
\alias{get_stat_fns}
\title{Build ggplot stat_function layers for each pseudolikelihood}
\usage{
get_stat_fns(pseudolikelihoods, LL_df)
}
\arguments{
\item{pseudolikelihoods}{List of functions.}

\item{LL_df}{Data frame containing psi grid.}
}
\value{
A named list of ggplot2 stat layers.
}
\description{
Build ggplot stat_function layers for each pseudolikelihood
}
\keyword{internal}
