% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-system.R
\name{has_comparison_result}
\alias{has_comparison_result}
\title{Does the workspace hold a comparison result?}
\usage{
has_comparison_result(ws)
}
\description{
Does the workspace hold a comparison result?
}
