% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-system.R
\name{has_integrated_result}
\alias{has_integrated_result}
\title{Does the workspace hold an integrated likelihood result?}
\usage{
has_integrated_result(ws)
}
\description{
Does the workspace hold an integrated likelihood result?
}
