% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-system.R
\name{has_profile_result}
\alias{has_profile_result}
\title{Does the workspace hold a profile likelihood result?}
\usage{
has_profile_result(ws)
}
\description{
Does the workspace hold a profile likelihood result?
}
