% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-system.R
\name{is_integrated}
\alias{is_integrated}
\title{Has integrate() been run on this model?
NOTE: "integrated" also appears as a result type; is_model() disambiguates.}
\usage{
is_integrated(x)
}
\description{
Has integrate() been run on this model?
NOTE: "integrated" also appears as a result type; is_model() disambiguates.
}
