% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-system.R
\name{is_integrated_result}
\alias{is_integrated_result}
\title{Is this an integrated likelihood result?
NOTE: "integrated" also appears as a model state marker; is_result() disambiguates.}
\usage{
is_integrated_result(x)
}
\description{
Is this an integrated likelihood result?
NOTE: "integrated" also appears as a model state marker; is_result() disambiguates.
}
