% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{make_ci_hline_layer}
\alias{make_ci_hline_layer}
\title{Horizontal CI cutoff lines (multi-curve)}
\usage{
make_ci_hline_layer(crit_df)
}
\description{
Horizontal CI cutoff lines (multi-curve)
}
\keyword{internal}
