% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tables.R
\name{make_kable_caption}
\alias{make_kable_caption}
\title{Construct Caption for Kable Tables}
\usage{
make_kable_caption(path)
}
\arguments{
\item{path}{Character path containing experiment identifiers.}
}
\value{
HTML string for use as caption.
}
\description{
Construct Caption for Kable Tables
}
