% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{make_label_repel}
\alias{make_label_repel}
\title{Repelled labels for point annotations}
\usage{
make_label_repel(label_data, y)
}
\description{
Repelled labels for point annotations
}
\keyword{internal}
