% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{make_label_vlines}
\alias{make_label_vlines}
\title{Vertical reference lines for labeled points}
\usage{
make_label_vlines(label_data, comparison = FALSE)
}
\arguments{
\item{label_data}{Data frame with columns source, value}

\item{comparison}{Logical; use comparison styling?}
}
\description{
Vertical reference lines for labeled points
}
\keyword{internal}
