% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval-omega-hat.R
\name{make_omega_hat_sampler}
\alias{make_omega_hat_sampler}
\title{Construct ω̂ Sampler for ψ(ω̂) = ψ_MLE}
\usage{
make_omega_hat_sampler(cal)
}
\arguments{
\item{cal}{A \code{likelyr_calibrated} model object.}
}
\value{
A function \code{f(init_guess)} returning ω̂.
}
\description{
Solves the manifold equation:
ψ(ω̂) - ψ_MLE = 0
using \code{nloptr::auglag()} and incorporating:

• θ_lower / θ_upper bounds
• inequality constraints h(θ) ≤ 0
• Jacobians for constraints (if supplied)
• optimizer_spec settings only for SLSQP/local solver
}
