% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{make_point_estimate_layer}
\alias{make_point_estimate_layer}
\title{Point estimate marker (single-likelihood plots)}
\usage{
make_point_estimate_layer(x, y, pseudolikelihood)
}
\description{
Point estimate marker (single-likelihood plots)
}
\keyword{internal}
