% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{make_stat_fn}
\alias{make_stat_fn}
\title{Build stat_function layer for pseudo-log-likelihood curves}
\usage{
make_stat_fn(
  psi_endpoints,
  zero_max_psi_ll_fn,
  pseudolikelihood,
  comparison = FALSE
)
}
\arguments{
\item{psi_endpoints}{Numeric range of psi.}

\item{zero_max_psi_ll_fn}{Zero-shifted log-likelihood function.}

\item{pseudolikelihood}{One of "integrated", "profile".}

\item{comparison}{Logical; use comparison styling?}
}
\description{
Build stat_function layer for pseudo-log-likelihood curves
}
\keyword{internal}
