% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-system.R
\name{mark_compared}
\alias{mark_compared}
\title{Mark a workspace as compared}
\usage{
mark_compared(x)
}
\arguments{
\item{x}{A workspace object.}
}
\value{
The same object with class \code{"compared"} prepended.
}
\description{
Mark a workspace as compared
}
\keyword{internal}
