% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-system.R
\name{mark_diagnosed}
\alias{mark_diagnosed}
\title{Mark a result object as having diagnostics computed}
\usage{
mark_diagnosed(x)
}
\description{
Mark a result object as having diagnostics computed
}
\keyword{internal}
