% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-system.R
\name{mark_inferred}
\alias{mark_inferred}
\title{Mark a result object as having inference computed}
\usage{
mark_inferred(x)
}
\description{
Mark a result object as having inference computed
}
\keyword{internal}
