% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-system.R
\name{mark_integrated}
\alias{mark_integrated}
\title{Mark a calibrated model as integrated}
\usage{
mark_integrated(x)
}
\arguments{
\item{x}{A calibrated model object.}
}
\value{
The same object with class \code{"integrated"} prepended.
}
\description{
Mark a calibrated model as integrated
}
\keyword{internal}
