% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-system.R
\name{mark_preprocessed}
\alias{mark_preprocessed}
\title{Mark a model as having been preprocessed}
\usage{
mark_preprocessed(x)
}
\description{
Mark a model as having been preprocessed
}
\keyword{internal}
