% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-system.R
\name{mark_profiled}
\alias{mark_profiled}
\title{Mark a calibrated model as profiled}
\usage{
mark_profiled(x)
}
\arguments{
\item{x}{A calibrated model object.}
}
\value{
The same object with class \code{"profiled"} prepended.
}
\description{
Mark a calibrated model as profiled
}
\keyword{internal}
