% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-branches.R
\name{plot_branches}
\alias{plot_branches}
\title{Plot Integrated Log-Likelihood Branches}
\usage{
plot_branches(x)
}
\arguments{
\item{mat}{Branch matrix.}

\item{crit}{Critical value for horizontal cutoff line.}
}
\value{
A ggplot object.
}
\description{
Plot Integrated Log-Likelihood Branches
}
