% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-style.R
\name{plot_ci_palette}
\alias{plot_ci_palette}
\title{Generate CI color palette}
\usage{
plot_ci_palette(interval_estimate_df)
}
\arguments{
\item{interval_estimate_df}{Data frame containing interval metadata.}
}
\value{
Named character vector of colors.
}
\description{
Generate CI color palette
}
