% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-style.R
\name{plot_curve_color}
\alias{plot_curve_color}
\title{Curve color}
\usage{
plot_curve_color(pseudolikelihood, comparison = FALSE)
}
\arguments{
\item{pseudolikelihood}{"integrate" or "profile"}

\item{comparison}{Logical}
}
\value{
Hex color
}
\description{
Curve color
}
