% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-style.R
\name{plot_curve_linetype}
\alias{plot_curve_linetype}
\title{Curve linetype}
\usage{
plot_curve_linetype(pseudolikelihood, comparison = FALSE)
}
\arguments{
\item{pseudolikelihood}{"integrate" or "profile"}

\item{comparison}{Logical}
}
\value{
Character linetype
}
\description{
Curve linetype
}
