% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-style.R
\name{plot_diagnostics_method_cfg}
\alias{plot_diagnostics_method_cfg}
\title{Diagnostics method override}
\usage{
plot_diagnostics_method_cfg(method)
}
\arguments{
\item{method}{"integrate" or "profile"}
}
\description{
Diagnostics method override
}
