% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-style.R
\name{plot_diagnostics_plot_cfg}
\alias{plot_diagnostics_plot_cfg}
\title{Diagnostics plot override}
\usage{
plot_diagnostics_plot_cfg(plot_name)
}
\arguments{
\item{plot_name}{Plot key}
}
\description{
Diagnostics plot override
}
