% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-style.R
\name{plot_diagnostics_style}
\alias{plot_diagnostics_style}
\title{Resolve diagnostics style}
\usage{
plot_diagnostics_style(method, plot)
}
\arguments{
\item{method}{Method}

\item{plot}{Plot name}
}
\value{
Named list
}
\description{
Resolve diagnostics style
}
