% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-style.R
\name{plot_point_estimate_color}
\alias{plot_point_estimate_color}
\title{Point estimate color}
\usage{
plot_point_estimate_color(pseudolikelihood, comparison = FALSE)
}
\arguments{
\item{pseudolikelihood}{"integrate" or "profile"}

\item{comparison}{Logical}
}
\value{
Hex color
}
\description{
Point estimate color
}
