% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-style.R
\name{plot_point_estimate_linetype}
\alias{plot_point_estimate_linetype}
\title{Point estimate linetype}
\usage{
plot_point_estimate_linetype(pseudolikelihood, comparison = FALSE)
}
\arguments{
\item{pseudolikelihood}{"integrate" or "profile"}

\item{comparison}{Logical}
}
\value{
Character linetype
}
\description{
Point estimate linetype
}
