% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-style.R
\name{plot_point_estimate_size}
\alias{plot_point_estimate_size}
\title{Get point estimate marker size}
\usage{
plot_point_estimate_size()
}
\value{
Numeric size
}
\description{
Get point estimate marker size
}
