% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-likelihood.R
\name{plot_pseudolikelihood_curve}
\alias{plot_pseudolikelihood_curve}
\title{Plot pseudolikelihood curve with inference annotations}
\usage{
plot_pseudolikelihood_curve(
  psi_ll_df,
  zero_max_psi_ll_fn,
  point_estimate_df,
  interval_estimate_df
)
}
\arguments{
\item{psi_ll_df}{Data frame of evaluated log-likelihood values}

\item{zero_max_psi_ll_fn}{Zero-shifted log-likelihood function}

\item{point_estimate_df}{Data frame with psi_hat and psi_0}

\item{interval_estimate_df}{Data frame of interval estimates}
}
\value{
ggplot object (invisibly)
}
\description{
Plot pseudolikelihood curve with inference annotations
}
\keyword{internal}
