% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-likelihood.R
\name{plot_pseudolikelihood_curves}
\alias{plot_pseudolikelihood_curves}
\title{Compare multiple pseudolikelihood curves}
\usage{
plot_pseudolikelihood_curves(res_list)
}
\arguments{
\item{res_list}{Named list of result objects}
}
\value{
ggplot object (invisibly)
}
\description{
Compare multiple pseudolikelihood curves
}
\keyword{internal}
