% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-likelihood.R
\name{plot_pseudolikelihood_points}
\alias{plot_pseudolikelihood_points}
\title{Plot pseudolikelihood evaluation points}
\usage{
plot_pseudolikelihood_points(psi_ll_df)
}
\arguments{
\item{psi_ll_df}{Data frame with columns psi, loglik and attribute "type"}
}
\value{
ggplot object (invisibly)
}
\description{
Plot pseudolikelihood evaluation points
}
\keyword{internal}
