% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-style.R
\name{plot_reference_line_style}
\alias{plot_reference_line_style}
\title{Reference line style}
\usage{
plot_reference_line_style(name)
}
\arguments{
\item{name}{Style key}
}
\value{
Named list
}
\description{
Reference line style
}
