% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-style.R
\name{plot_theme_cfg}
\alias{plot_theme_cfg}
\title{Get full ggplot theme configuration}
\usage{
plot_theme_cfg()
}
\value{
Named list corresponding to styles/plots.yml$theme
}
\description{
Get full ggplot theme configuration
}
