% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-nuisance.R
\name{print.nuisance_spec}
\alias{print.nuisance_spec}
\title{Print Method for \code{nuisance_spec}}
\usage{
\method{print}{nuisance_spec}(x, ...)
}
\arguments{
\item{x}{A \code{nuisance_spec} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Displays a concise summary of the nuisance specification.
}
