% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer-render.R
\name{render_estimate_table}
\alias{render_estimate_table}
\title{Render combined estimate table}
\usage{
render_estimate_table(estimate_df)
}
\arguments{
\item{estimate_df}{Data frame combining point and interval estimates.}
}
\value{
A kableExtra object.
}
\description{
Render combined estimate table
}
\keyword{internal}
