% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer-render.R
\name{render_interval_estimate_kable}
\alias{render_interval_estimate_kable}
\title{Render confidence interval table}
\usage{
render_interval_estimate_kable(interval_estimate_df, show_caption = TRUE)
}
\arguments{
\item{interval_estimate_df}{Data frame produced by
\code{get_interval_estimate_df()}.}

\item{show_caption}{Logical; whether to display a caption.}
}
\value{
A kableExtra object.
}
\description{
Render confidence interval table
}
