% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer-render.R
\name{render_interval_estimate_table}
\alias{render_interval_estimate_table}
\title{Render confidence interval table}
\usage{
render_interval_estimate_table(interval_estimate_df)
}
\arguments{
\item{interval_estimate_df}{Data frame produced by
\code{get_interval_estimate_df()}.}
}
\value{
A kableExtra object.
}
\description{
Render confidence interval table
}
\keyword{internal}
