% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-render.R, R/infer-render.R
\name{render_interval_estimates_kable}
\alias{render_interval_estimates_kable}
\title{Render confidence interval table}
\usage{
render_interval_estimates_kable(interval_estimates_df, show_caption = TRUE)

render_interval_estimates_kable(interval_estimates_df, show_caption = TRUE)
}
\arguments{
\item{interval_estimates_df}{Data frame produced by
\code{get_interval_estimates_df()}.}

\item{show_caption}{Logical; whether to display a caption.}
}
\value{
A kableExtra object.

A kableExtra object.
}
\description{
Render confidence interval table

Render confidence interval table
}
