% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tables.R
\name{render_LL_comparison_tables}
\alias{render_LL_comparison_tables}
\title{Render Comparison Tables for MLEs and Confidence Intervals}
\usage{
render_LL_comparison_tables(MLE_data, conf_ints, psi_0)
}
\arguments{
\item{MLE_data}{Data frame with MLEs per pseudolikelihood.}

\item{conf_ints}{Confidence interval table.}

\item{psi_0}{True ψ value (optional).}
}
\value{
A list of two HTML tables.
}
\description{
Render Comparison Tables for MLEs and Confidence Intervals
}
