% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-render.R
\name{render_point_estimate_comparison}
\alias{render_point_estimate_comparison}
\title{Render point estimate comparison table}
\usage{
render_point_estimate_comparison(res)
}
\arguments{
\item{res}{}
}
\value{
A kableExtra object.
}
\description{
Render point estimate comparison table
}
