% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer-render.R
\name{render_point_estimate_kable}
\alias{render_point_estimate_kable}
\title{Render point estimate table}
\usage{
render_point_estimate_kable(point_estimate_df, show_caption = TRUE)
}
\arguments{
\item{point_estimate_df}{Data frame with columns \code{psi_0}, \code{psi_hat},
and \code{se_psi_hat}.}

\item{show_caption}{Logical; whether to display a caption.}
}
\value{
A kableExtra object.
}
\description{
Render point estimate table
}
