% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch-utils.R
\name{snap_to_grid}
\alias{snap_to_grid}
\title{Snap ψ to Nearest Grid Point}
\usage{
snap_to_grid(psi, grid)
}
\arguments{
\item{psi}{Numeric scalar.}

\item{grid}{A \code{"psi_grid"} object.}
}
\value{
Numeric: nearest ψ-grid point.
}
\description{
Given a ψ value and a ψ-grid created by \code{\link[=psi_grid_anchor]{psi_grid_anchor()}},
return the nearest ψ-grid point.
}
\keyword{internal}
