% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ci.R
\name{summarize_mle_performance}
\alias{summarize_mle_performance}
\title{Summarize MLE Performance Across Simulations}
\usage{
summarize_mle_performance(mle_list, true_value)
}
\arguments{
\item{mle_list}{A list where each element contains a tibble with an MLE column.}

\item{true_value}{The true ψ used for bias/SD/RMSE.}
}
\value{
A tibble summarizing bias, SD, and RMSE.
}
\description{
Summarize MLE Performance Across Simulations
}
