% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-palette.R
\name{table_accent}
\alias{table_accent}
\title{Get table accent color}
\usage{
table_accent(key)
}
\arguments{
\item{key}{One of "separator", "header_row", "group_row"}
}
\description{
Get table accent color
}
