% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-palette.R
\name{table_body_bg}
\alias{table_body_bg}
\title{Get default table body background}
\usage{
table_body_bg(section)
}
\arguments{
\item{section}{One of "point_estimate", "interval_estimate", "truth"}
}
\description{
Get default table body background
}
