% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-palette.R
\name{table_column_bg}
\alias{table_column_bg}
\title{Get semantic column background color}
\usage{
table_column_bg(key)
}
\arguments{
\item{key}{One of "point", "truth", "interval", "diagram"}
}
\description{
Get semantic column background color
}
