% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-palette.R
\name{table_column_header_bg}
\alias{table_column_header_bg}
\title{Get column header background color}
\usage{
table_column_header_bg(section)
}
\arguments{
\item{section}{One of "point_estimate", "interval_estimate", "truth"}
}
\description{
Get column header background color
}
