% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-palette.R
\name{table_group_header_bg}
\alias{table_group_header_bg}
\title{Get group header background color}
\usage{
table_group_header_bg(group)
}
\arguments{
\item{group}{One of "point_estimate", "interval_estimate", "truth"}
}
\description{
Get group header background color
}
