% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-palette.R
\name{table_row_block_bg}
\alias{table_row_block_bg}
\title{Get alternating row block background}
\usage{
table_row_block_bg(index, type = c("level", "pseudolikelihood"))
}
\arguments{
\item{index}{Integer index (1-based)}

\item{type}{One of "pseudolikelihood", "level"}
}
\description{
Get alternating row block background
}
