% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-palette.R
\name{table_text_body}
\alias{table_text_body}
\title{Get table body text color by semantic name}
\usage{
table_text_body(key)
}
\arguments{
\item{key}{One of the names in tables.yml$text$body}
}
\value{
Hex color string
}
\description{
Get table body text color by semantic name
}
