% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-palette.R
\name{table_text_header}
\alias{table_text_header}
\title{Get table header text color}
\usage{
table_text_header(key = c("column", "group"))
}
\arguments{
\item{key}{One of "column", "group"}
}
\description{
Get table header text color
}
