% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-execution.R
\name{total_branches}
\alias{total_branches}
\title{Total Number of Monte Carlo Branches}
\usage{
total_branches(x)
}
\arguments{
\item{x}{An execution_spec.}
}
\value{
Integer number of branches.
}
\description{
Returns the total number of Monte Carlo branches implied by an
execution_spec (serial or parallel).
}
