% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view}
\alias{view}
\title{View an object}
\usage{
view(x, ...)
}
\arguments{
\item{x}{Object to view.}

\item{...}{Passed to methods.}
}
\description{
Render the canonical human-facing representation of an object.

Unlike \code{View()}, this is a semantic viewer that may print tables,
plots, or structured summaries.
}
