% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer-interval_estimate.R
\name{add_interval_diagnostics}
\alias{add_interval_diagnostics}
\title{Add Diagnostic Quantities to Psi Confidence Intervals}
\usage{
add_interval_diagnostics(interval_estimate_df, psi_0 = NA_real_)
}
\arguments{
\item{interval_estimate_df}{Data frame with columns \code{alpha},
\code{lower}, \code{upper}.}

\item{psi_0}{Optional numeric scalar. True value of \eqn{\psi}.}
}
\value{
Data frame with additional diagnostic columns.
}
\description{
Add Diagnostic Quantities to Psi Confidence Intervals
}
\keyword{internal}
