% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch-generate.R
\name{build_one_branch}
\alias{build_one_branch}
\title{Build a full branch (left + right) around a branch-specific mode}
\usage{
build_one_branch(
  psi_MLE,
  increment,
  left_start,
  right_start,
  branch_cutoff,
  init_guess,
  max_retries,
  eval_psi_fun
)
}
\description{
Build a full branch (left + right) around a branch-specific mode
}
\keyword{internal}
