% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate a Likelihood Workflow}
\usage{
calibrate(wf, data, type = c("IL", "PL"), verbose = TRUE)
}
\arguments{
\item{wf}{A \code{likelihood_workflow} object created by \code{\link[=workflow]{workflow()}}.}

\item{data}{User data to embed into model/estimand closures.}

\item{type}{\code{"IL"} or \code{"PL"}.}

\item{verbose}{Logical; prints diagnostics if TRUE.}
}
\value{
S3 object of class:
\itemize{
\item \verb{"IL_calibration", "likelihood_calibration"}
\item \verb{"PL_calibration", "likelihood_calibration"}
}
}
\description{
Calibrates a likelihood workflow object for either integrated likelihood
(IL) or profile likelihood (PL) generation. This embeds data into all
likelihood & estimand functions, computes \eqn{\theta_{\mathrm{MLE}}},
validates the search interval, and attaches IL- or PL-specific fields.
}
