% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate-estimand.R
\name{calibrate_estimand}
\alias{calibrate_estimand}
\title{Calibrate Estimand Component}
\usage{
calibrate_estimand(estimand, data, theta_mle)
}
\arguments{
\item{estimand}{An estimand_spec object.}

\item{data}{User data.}

\item{theta_mle}{Numeric MLE of theta.}
}
\value{
The SAME estimand_spec object, enriched with:
• $psi_fn           (data-bound)
• $psi_jac          (data-bound, if applicable)
• $psi_mle
• $search_interval
}
\description{
Produces data-bound psi_fn(), computes psi_MLE and search interval.
}
