% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate-likelihood.R
\name{calibrate_likelihood}
\alias{calibrate_likelihood}
\title{Calibrate Likelihood Component}
\usage{
calibrate_likelihood(likelihood, data)
}
\arguments{
\item{likelihood}{A likelihood_spec object.}

\item{data}{User data.}
}
\value{
The SAME likelihood_spec object, enriched with:
• $loglik     — data-bound closure
• $theta_mle  — analytic MLE computed on \code{data}
}
\description{
Binds data to the likelihood functions and computes theta_MLE.
}
