% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate-nuisance.R
\name{calibrate_nuisance}
\alias{calibrate_nuisance}
\title{Calibrate Nuisance Component}
\usage{
calibrate_nuisance(nuisance, data)
}
\arguments{
\item{nuisance}{A nuisance_spec object or NULL.}

\item{data}{User data.}
}
\value{
The SAME nuisance_spec object, with:
• $E_loglik       (data-bound)
• $E_loglik_grad  (data-bound, if present)
}
\description{
Calibrate Nuisance Component
}
