% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-branch-params.R
\name{compute_num_branches}
\alias{compute_num_branches}
\title{Compute number of Monte Carlo branches (R)}
\usage{
compute_num_branches(execution)
}
\arguments{
\item{execution}{An execution spec object created by \code{\link[=serial_spec]{serial_spec()}} or \code{\link[=parallel_spec]{parallel_spec()}}}
}
\value{
Integer number of branches (R)
}
\description{
Determines how many nuisance-based branches should be drawn for
Monte Carlo integrated likelihood, based on the execution specification.

For:
\itemize{
\item serial:   R is explicitly provided via \code{serial_spec(R = ...)}
\item parallel: R = num_workers * chunk_size
}
}
\keyword{internal}
