% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch-generate.R
\name{generate_branches}
\alias{generate_branches}
\title{Generate Monte Carlo Branches (Integrated Log-Likelihood Only)}
\usage{
generate_branches(cal, verbose = TRUE)
}
\description{
Generate Monte Carlo Branches (Integrated Log-Likelihood Only)
}
\keyword{internal}
