% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-infer.R
\name{infer}
\alias{infer}
\title{Likelihood-Based Inference for a Calibrated Model}
\usage{
infer(cal, which = NULL, alpha_levels = NULL)
}
\description{
Likelihood-Based Inference for a Calibrated Model
}
