% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-integrate.R
\name{integrate}
\alias{integrate}
\title{Integrated Likelihood}
\usage{
integrate(cal, ...)
}
\arguments{
\item{cal}{A calibrated_model produced by \code{calibrate()}.}

\item{...}{Additional arguments.}

\item{verbose}{Logical.}
}
\value{
An \code{integrated_likelihood} object.
}
\description{
Computes the integrated likelihood (IL) after calibration with
\code{calibrate()}. The method attaches IL-specific components to \code{cal}
and then runs the Monte Carlo branch engine.
}
