% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-likelihood.R
\name{likelihood_spec}
\alias{likelihood_spec}
\title{Specify a Parametric Likelihood Model}
\usage{
likelihood_spec(
  loglik,
  theta_dim,
  theta_mle_fn,
  theta_lower = NULL,
  theta_upper = NULL,
  ineq = NULL,
  ineq_jac = NULL,
  name = NULL,
  ...
)
}
\value{
A \code{likelihood_spec} object with classes
c("likelihood_spec", "likelyr_spec").
}
\description{
Defines:
• log-likelihood p(y | theta)
• dimension J of theta
• analytic theta_MLE initializer
• model-family constraints: box + inequality
}
