% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-primitives.R
\name{make_stat_fn}
\alias{make_stat_fn}
\title{Build ggplot stat_function layer for pseudolikelihood curves}
\usage{
make_stat_fn(psi_endpoints, relative_loglik_fn)
}
\arguments{
\item{psi_endpoints}{Numeric range.}

\item{relative_loglik_fn}{Function of psi.}
}
\value{
A ggplot stat_function layer.
}
\description{
Build ggplot stat_function layer for pseudolikelihood curves
}
\keyword{internal}
