% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-system.R
\name{mark_diagnosed}
\alias{mark_diagnosed}
\title{Mark a result as diagnosed}
\usage{
mark_diagnosed(x)
}
\arguments{
\item{x}{A result object.}
}
\value{
The same object with class \code{"diagnostics"} prepended.
}
\description{
Mark a result as diagnosed
}
\keyword{internal}
