% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-system.R
\name{mark_inferred}
\alias{mark_inferred}
\title{Mark a result as inferred}
\usage{
mark_inferred(x)
}
\arguments{
\item{x}{A result object.}
}
\value{
The same object with class \code{"inference"} prepended.
}
\description{
Mark a result as inferred
}
\keyword{internal}
