% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-system.R
\name{mark_profiled}
\alias{mark_profiled}
\title{Mark a model as having a profile likelihood computed}
\usage{
mark_profiled(x)
}
\description{
Mark a model as having a profile likelihood computed
}
\keyword{internal}
