% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-execution.R
\name{parallel_spec}
\alias{parallel_spec}
\title{Parallel Execution Specification (foreach + \%dofuture\%)}
\usage{
parallel_spec(
  num_workers,
  chunk_size = 1L,
  seed = TRUE,
  packages = character(),
  name = "parallel"
)
}
\value{
A \code{parallel_spec} object.
}
\description{
Parallel Execution Specification (foreach + \%dofuture\%)
}
