% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{plot_base}
\alias{plot_base}
\title{Base Dark Theme for Log-Likelihood Plots}
\usage{
plot_base()
}
\value{
A ggplot2 object representing the base plotting surface.
}
\description{
Base Dark Theme for Log-Likelihood Plots
}
\keyword{internal}
