% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-style.R
\name{plot_curve_color}
\alias{plot_curve_color}
\title{Get curve color by pseudolikelihood}
\usage{
plot_curve_color(pseudolikelihood, comparison = FALSE)
}
\arguments{
\item{pseudolikelihood}{One of "integrated", "profile"}

\item{comparison}{Logical; use comparison styling?}
}
\value{
Hex color string
}
\description{
Get curve color by pseudolikelihood
}
