% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-style.R
\name{plot_curve_linetype}
\alias{plot_curve_linetype}
\title{Get curve linetype}
\usage{
plot_curve_linetype(pseudolikelihood, comparison = FALSE)
}
\arguments{
\item{pseudolikelihood}{One of "integrated", "profile"}

\item{comparison}{Logical; use comparison styling?}
}
\value{
Character linetype
}
\description{
Get curve linetype
}
