% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-style.R
\name{plot_point_estimate_color}
\alias{plot_point_estimate_color}
\title{Get point estimate color by pseudolikelihood}
\usage{
plot_point_estimate_color(pseudolikelihood, comparison = FALSE)
}
\arguments{
\item{pseudolikelihood}{One of "integrated", "profile"}

\item{comparison}{Logical; use comparison styling?}
}
\value{
Hex color string
}
\description{
Get point estimate color by pseudolikelihood
}
