% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-style.R
\name{plot_point_estimate_linetype}
\alias{plot_point_estimate_linetype}
\title{Get point estimate linetype}
\usage{
plot_point_estimate_linetype(pseudolikelihood, comparison = FALSE)
}
\arguments{
\item{pseudolikelihood}{One of "integrated", "profile"}

\item{comparison}{Logical; use comparison styling?}
}
\value{
Character linetype
}
\description{
Get point estimate linetype
}
