% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-profile.R
\name{profile}
\alias{profile}
\title{Profile Likelihood}
\usage{
profile(cal, ...)
}
\arguments{
\item{cal}{A calibrated_model.}
}
\description{
Placeholder function. The profile likelihood engine is not yet
implemented. Calling this function will throw an informative error.
}
