% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-execution.R
\name{serial_spec}
\alias{serial_spec}
\title{Serial Execution Specification}
\usage{
serial_spec(packages = character(), name = "serial")
}
\arguments{
\item{packages}{Character vector of packages to load.}

\item{name}{Optional identifier.}
}
\value{
A \code{serial_spec} object.
}
\description{
Runs Monte Carlo branch generation \strong{in serial} (no parallelism).
}
