% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-execution.R
\name{serial_spec}
\alias{serial_spec}
\title{Serial Execution Specification}
\usage{
serial_spec(R, seed = TRUE, packages = character(), name = "serial")
}
\value{
A \code{serial_spec} object.
}
\description{
Runs Monte Carlo computation \strong{in serial}. No parallel workers.
}
