% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer-interval_estimates.R
\name{shift_psi_ll_fn}
\alias{shift_psi_ll_fn}
\title{Shift a Psi Log-Likelihood Function}
\usage{
shift_psi_ll_fn(psi_ll_fn, shift_val)
}
\arguments{
\item{shift_val}{Numeric scalar giving the shift value.}

\item{zero_max_psi_ll_fn}{Function returning \eqn{\ell(\psi)}.}
}
\value{
A function of \eqn{\psi} returning the shifted log-likelihood.
}
\description{
Returns a new function corresponding to
\eqn{\ell(\psi) - c}, where \eqn{c} is a constant shift.
}
\keyword{internal}
