% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch-traverse.R
\name{traverse_branch_side}
\alias{traverse_branch_side}
\title{One-Sided Branch Sweep Along the ψ-Grid}
\usage{
traverse_branch_side(
  grid,
  k_direction,
  k_start,
  branch_cutoff,
  init_guess,
  branch_evaluator,
  max_retries = 0L,
  stop_at_bounds = TRUE,
  eval_at_bounds = TRUE
)
}
\description{
One-Sided Branch Sweep Along the ψ-Grid
}
\keyword{internal}
