% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow}
\alias{workflow}
\title{Create a likelihood workflow}
\usage{
workflow(
  model = NULL,
  estimand = NULL,
  nuisance = NULL,
  optimizer = NULL,
  execution = NULL
)
}
\arguments{
\item{model}{Optional object created by \code{\link[=model_spec]{model_spec()}}}

\item{estimand}{Optional object created by \code{\link[=estimand_spec]{estimand_spec()}}}

\item{nuisance}{Optional object created by \code{\link[=nuisance_spec]{nuisance_spec()}}}

\item{optimizer}{Optional object created by \code{\link[=optimizer_spec]{optimizer_spec()}}}

\item{execution}{Optional object created by \code{\link[=execution_spec]{execution_spec()}}}
}
\value{
An S3 object of class \code{likelihood_workflow}
}
\description{
A workflow stores the configuration required to compute profile or
integrated (pseudo-)likelihood function values for a statistical model.

You may construct it all at once:

\code{workflow(model, estimand, nuisance, optimizer, execution)}

or incrementally using a pipe:

\code{workflow() |> add(model_spec) |> add(estimand_spec) |> ...}

No estimation is performed here — the workflow is declarative.
}
